kstats: implement circular logs, more config options, build fixes
Add the ability to collect timestamps (or samples) as a circular log,
and reduce the size of the table.

A new function kstats2_new() support more config options including the
use of the collector as a circular log. Samples (individual values, with
optional timestamps) are stored in a log, the log can stop when full or
wrap around.
Print routines and manual configuration via debugfs are updated accordingly,
as an example:

    echo "trace pcpu:find_vma entries 1600 wrap 1" > _control

On passing, fix build on linux 5.11 and above